+++ src/connectedk8s/azext_connectedk8s/custom.py
@@ -81,7 +81,7 @@ if TYPE_CHECKING:
     from azure.cli.core.commands import AzCliCommand
     from azure.core.polling import LROPoller
     from Crypto.PublicKey.RSA import RsaKey
-    from knack.commands import CLICommmand
+    from knack.commands import CLICommand
     from kubernetes.client import V1NodeList
     from kubernetes.config.kube_config import ConfigNode
     from requests.models import Response
@@ -99,7 +99,7 @@ logger = get_logger(__name__)
 
 
 def create_connectedk8s(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -235,6 +235,10 @@ def create_connectedk8s(
         configuration_settings,
         configuration_protected_settings,
     )
+
+    # Validate and update bundle feature flag value if provided
+    validate_connect_cluster_bundle_feature_flag_value(configuration_settings, yes)
+
     arc_agentry_configurations = generate_arc_agent_configuration(
         configuration_settings, configuration_protected_settings
     )
@@ -1043,7 +1047,7 @@ def validate_existing_provisioned_cluster_for_reput(
                 raise InvalidArgumentValueError(err_msg)
 
 
-def send_cloud_telemetry(cmd: CLICommmand) -> str:
+def send_cloud_telemetry(cmd: CLICommand) -> str:
     telemetry.add_extension_event(
         "connectedk8s", {"Context.Default.AzureCLI.AzureCloud": cmd.cli_ctx.cloud.name}
     )
@@ -1289,7 +1293,7 @@ def connected_cluster_exists(
     return True
 
 
-def get_default_config_dp_endpoint(cmd: CLICommmand, location: str) -> str:
+def get_default_config_dp_endpoint(cmd: CLICommand, location: str) -> str:
     cloud_based_domain = cmd.cli_ctx.cloud.endpoints.active_directory.split(".")[2]
     config_dp_endpoint = (
         f"https://{location}.dp.kubernetesconfiguration.azure.{cloud_based_domain}"
@@ -1298,7 +1302,7 @@ def get_default_config_dp_endpoint(cmd: CLICommmand, location: str) -> str:
 
 
 def get_config_dp_endpoint(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     location: str,
     values_file: str | None,
     arm_metadata: dict[str, Any] | None = None,
@@ -1510,6 +1514,169 @@ def set_security_profile(enable_workload_identity: bool) -> SecurityProfile:
         )
     )
     return security_profile
+
+
+def get_k8s_extension_module(module_name):
+    try:
+        # adding the installed extension in the path
+        from azure.cli.core.extension.operations import add_extension_to_path
+        add_extension_to_path(consts.CONST_K8S_EXTENSION_NAME)
+        # import the extension module
+        from importlib import import_module
+        azext_custom = import_module(module_name)
+        return azext_custom
+    except ImportError:
+        raise ImportError(  # pylint: disable=raise-missing-from
+            "Please add CLI extension `k8s-extension` to disable bundle feature flag.\n"
+            "Run command `az extension add --name k8s-extension`"
+        )
+
+
+def get_bundle_feature_flag_from_arc_agentry_config(
+    current_arc_agentry_config: list[ArcAgentryConfigurations]
+) -> str | None:
+    for agentry_config in current_arc_agentry_config:
+        if agentry_config.feature == consts.Arc_Agentry_Bundle_Feature and \
+           consts.Arc_Agentry_Bundle_Feature_Setting in agentry_config.settings:
+            return agentry_config.settings[consts.Arc_Agentry_Bundle_Feature_Setting].lower()
+    return None
+
+
+def get_bundle_feature_flag_from_configuration_settings(
+    configuration_settings: dict[str, Any]
+) -> str | None:
+    settings = configuration_settings.get(consts.Arc_Agentry_Bundle_Feature, {})
+    value = settings.get(consts.Arc_Agentry_Bundle_Feature_Setting)
+    return value if value is None else value.lower()
+
+
+def validate_bundle_feature_flag_value(
+    configuration_settings: dict[str, Any]
+) -> str | None:
+    print(f"Step: {utils.get_utctimestring()}: Validating the bundle feature flag value")
+    value = get_bundle_feature_flag_from_configuration_settings(configuration_settings)
+
+    if value is not None:
+        # Remove leading and trailing whitespace and quotes
+        value = value.strip().strip("'\"")
+
+        if value and value not in consts.Bundle_Feature_Value_List:
+            raise InvalidArgumentValueError(
+                f"Not supported value for the feature flag '{consts.Arc_Agentry_Bundle_Feature_Setting}': "
+                f"'{value}'. Please specify a value from the list: {consts.Bundle_Feature_Value_List}."
+            )
+
+        configuration_settings[consts.Arc_Agentry_Bundle_Feature][consts.Arc_Agentry_Bundle_Feature_Setting] = value
+        print(f"Step: {utils.get_utctimestring()}: Setting the bundle feature flag value to '{value}'")
+
+    return value
+
+
+def validate_connect_cluster_bundle_feature_flag_value(
+    configuration_settings: dict[str, Any],
+    yes: bool = False,
+):
+    bundle_feature_flag_value = validate_bundle_feature_flag_value(configuration_settings)
+
+    # If the bundle feature flag value is None, skip the validation
+    if bundle_feature_flag_value is None:
+        return
+
+    if bundle_feature_flag_value == "preview":
+        confirmation_message = (
+            f"You are about to enter the 'preview' mode for {consts.Arc_Agentry_Bundle_Feature_Setting}. "
+            "In this mode, all SLA support will be discontinued, and the cluster will remain in 'preview' mode "
+            "until it is disconnected from Arc. Are you sure you want to proceed? "
+        )
+
+        utils.user_confirmation(confirmation_message, yes)
+
+        logger.warning(
+            "Entered %s 'preview' mode. All SLA support is discontinued, and the cluster will remain in 'preview' mode "
+            "until it is disconnected from Arc.",
+            consts.Arc_Agentry_Bundle_Feature_Setting
+        )
+
+    elif bundle_feature_flag_value == "disabled":
+        err_msg = (
+            f"{consts.Arc_Agentry_Bundle_Feature_Setting} 'disabled' mode can only be set using 'az connectedk8s update'. "
+            f"To keep the bundle feature flag off during cluster connection, remove "
+            f"{consts.Arc_Agentry_Bundle_Feature_Setting} from the --config."
+        )
+        raise ArgumentUsageError(err_msg)
+
+
+def validate_update_cluster_bundle_feature_flag_value(
+    cmd: CLICommand,
+    current_arc_agentry_config: list[ArcAgentryConfigurations],
+    configuration_settings: dict[str, Any],
+    resource_group_name: str,
+    cluster_name: str,
+):
+    bundle_feature_flag_value = validate_bundle_feature_flag_value(configuration_settings)
+
+    # If the bundle feature flag value is None, skip the validation
+    if bundle_feature_flag_value is None:
+        return
+
+    current_bundle_feature_flag_value = get_bundle_feature_flag_from_arc_agentry_config(current_arc_agentry_config)
+
+    if bundle_feature_flag_value == "preview":
+        err_msg = (
+            f"{consts.Arc_Agentry_Bundle_Feature_Setting} 'preview' mode can only be enabled when a cluster "
+            "is first connected to Arc with 'az connectedk8s connect'. Updating the preview mode config with "
+            "'az connectedk8s update' is not allowed."
+        )
+        raise ArgumentUsageError(err_msg)
+
+    if current_bundle_feature_flag_value == "preview":
+        err_msg = (
+            f"The cluster is in {consts.Arc_Agentry_Bundle_Feature_Setting} 'preview' mode, "
+            "updating the value is not allowed."
+        )
+        raise ArgumentUsageError(err_msg)
+
+    invalid_transition = (
+        (current_bundle_feature_flag_value == "enabled" and bundle_feature_flag_value == "") or
+        (current_bundle_feature_flag_value == "" and bundle_feature_flag_value == "disabled")
+    )
+
+    if invalid_transition:
+        err_msg = (
+            f"Could not set {consts.Arc_Agentry_Bundle_Feature}.{consts.Arc_Agentry_Bundle_Feature_Setting} from "
+            f"'{current_bundle_feature_flag_value}' to '{bundle_feature_flag_value}'."
+        )
+
+        raise ArgumentUsageError(err_msg)
+
+    # If the bundle feature flag is set to 'disabled', check if any bundle extensions are installed
+    if current_bundle_feature_flag_value == "enabled" and bundle_feature_flag_value == "disabled":
+        client_factory = get_k8s_extension_module(consts.CONST_K8S_EXTENSION_CLIENT_FACTORY_MOD_NAME)
+        client = client_factory.cf_k8s_extension_operation(cmd.cli_ctx)
+        k8s_extension_custom_mod = get_k8s_extension_module(consts.CONST_K8S_EXTENSION_CUSTOM_MOD_NAME)
+        extensions = k8s_extension_custom_mod.list_k8s_extension(
+            client,
+            resource_group_name,
+            cluster_name,
+            "connectedClusters",
+        )
+
+        installed_bundle_extensions = [
+            ext.extension_type.lower()
+            for ext in extensions
+            if ext.extension_type.lower() in consts.Bundle_Extension_Type_List
+        ]
+
+        if installed_bundle_extensions:
+            err_msg = (
+                f"Could not set {consts.Arc_Agentry_Bundle_Feature}.{consts.Arc_Agentry_Bundle_Feature_Setting} to "
+                f"'disabled' - detected the following extension types on the cluster: {installed_bundle_extensions}.\n"
+                f"Please remove them with 'az k8s-extension delete --cluster-name <clusterName> "
+                f"--cluster-type <clusterType> --resource-group <resourceGroupName> --name <extensionName>' "
+                f"and try turning off the feature again."
+            )
+
+            raise ArgumentUsageError(err_msg)
 
 
 def generate_arc_agent_configuration(
@@ -1733,7 +1900,7 @@ def list_connectedk8s(
 
 
 def delete_connectedk8s(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -1995,7 +2162,7 @@ def update_connected_cluster_internal(
 
 
 def update_connected_cluster(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -2072,9 +2239,6 @@ def update_connected_cluster(
         configuration_settings,
         configuration_protected_settings,
     )
-    arc_agentry_configurations = generate_arc_agent_configuration(
-        configuration_settings, redacted_protected_values
-    )
 
     # Fetch Connected Cluster for agent version
     connected_cluster = client.get(resource_group_name, cluster_name)
@@ -2088,6 +2252,15 @@ def update_connected_cluster(
             + consts.Doc_Provisioned_Cluster_Update_Url
         )
         raise InvalidArgumentValueError(err_msg)
+
+    # Validate and update bundle feature flag value if provided
+    validate_update_cluster_bundle_feature_flag_value(
+        cmd, connected_cluster.arc_agentry_configurations, configuration_settings, resource_group_name, cluster_name
+    )
+
+    arc_agentry_configurations = generate_arc_agent_configuration(
+        configuration_settings, redacted_protected_values
+    )
 
     # Patching the connected cluster ARM resource
     arm_properties_unset = (
@@ -2351,7 +2524,7 @@ def update_connected_cluster(
 
 
 def upgrade_agents(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -2654,6 +2827,22 @@ def upgrade_agents(
             for message in consts.Helm_Install_Release_Userfault_Messages
         ):
             telemetry.set_user_fault()
+
+        namespace = "azure-arc"
+        label_selector = "job-name=agent-update-validator"
+
+        # Get the list of pods matching the label
+        pods = api_instance.list_namespaced_pod(namespace=namespace, label_selector=label_selector)
+
+        # Extract the terminated message from the container
+        if pods.items:
+            pod = pods.items[0]
+            container_statuses = pod.status.container_statuses
+            if container_statuses:
+                state = container_statuses[0].state
+                if state.terminated:
+                    helm_upgrade_error_message = state.terminated.message
+
         telemetry.set_exception(
             exception=error_helm_upgrade.decode("ascii"),
             fault_type=consts.Install_HelmRelease_Fault_Type,
@@ -2797,7 +2986,7 @@ def get_all_helm_values(
 
 
 def enable_features(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -3030,7 +3219,7 @@ def enable_features(
 
 
 def disable_features(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -3169,7 +3358,7 @@ def disable_features(
 
 
 def get_chart_and_disable_features(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     connected_cluster: ConnectedCluster,
     kube_config: str | None,
     kube_context: str | None,
@@ -3260,7 +3449,7 @@ def get_chart_and_disable_features(
 
 
 def disable_cluster_connect(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -3467,7 +3656,7 @@ def handle_merge(
 
 
 def client_side_proxy_wrapper(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,
@@ -3638,7 +3827,7 @@ def client_side_proxy_wrapper(
 
 
 def client_side_proxy_main(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     tenant_id: str,
     client: ConnectedClusterOperations,
     resource_group_name: str,
@@ -3709,7 +3898,7 @@ def client_side_proxy_main(
 
 
 def client_side_proxy(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     tenant_id: str,
     client: ConnectedClusterOperations,
     resource_group_name: str,
@@ -3842,7 +4031,7 @@ def client_side_proxy(
 
 
 def check_cl_registration_and_get_oid(
-    cmd: CLICommmand, cl_oid: str | None, subscription_id: str | None
+    cmd: CLICommand, cl_oid: str | None, subscription_id: str | None
 ) -> tuple[bool, str]:
     print(
         f"Step: {utils.get_utctimestring()}: Checking Custom Location(Microsoft.ExtendedLocation) RP Registration state for this Subscription, and attempt to get the Custom Location Object ID (OID),if registered"
@@ -3881,7 +4070,7 @@ def check_cl_registration_and_get_oid(
     return enable_custom_locations, custom_locations_oid
 
 
-def get_custom_locations_oid(cmd: CLICommmand, cl_oid: str | None) -> str:
+def get_custom_locations_oid(cmd: CLICommand, cl_oid: str | None) -> str:
     try:
         graph_client = graph_client_factory(cmd.cli_ctx)
         app_id = "bc313c14-388c-4e7d-a58e-70017303ee3b"
@@ -3942,7 +4131,7 @@ def get_custom_locations_oid(cmd: CLICommmand, cl_oid: str | None) -> str:
 
 
 def troubleshoot(
-    cmd: CLICommmand,
+    cmd: CLICommand,
     client: ConnectedClusterOperations,
     resource_group_name: str,
     cluster_name: str,